Split monolith app.js into angular component subfolders
