+++ client/controllers/customization.js
@@ -1,0 +1,4 @@
+app.controller('customizationController', function($scope, settings) {
+  $scope.settings = settings.settings;
+  $scope.settingControl = settings;
+});
