+++ client/controllers/dashboard.js
@@ -1,0 +1,6 @@
+app.controller('dashboardController', function($scope, dashboardSettings) {
+  dashboardSettings.setSettingsfromRouteParams();
+  $scope.noGroups = function() {
+    return dashboardSettings.numberGroups < 1;
+  };
+});
