+++ client/directives/status.js
@@ -1,0 +1,9 @@
+app.directive('status', function() {
+  return {
+    restrict: 'E',
+    templateUrl: 'partials/status.html',
+    scope: {
+      status: '='
+    }
+  };
+});
