+++ package.json
@@ -23,6 +23,7 @@
     "gulp-changed": "^1.2.1",
     "gulp-concat": "^2.5.2",
     "gulp-cssmin": "^0.1.7",
+    "gulp-insert": "^0.4.0",
     "gulp-less": "^3.0.3",
     "gulp-ng-annotate": "^1.0.0",
     "gulp-uglify": "^1.2.0",
